07: Create modal structure of pages and modal form. Impl save into DB

+++ src/main/kotlin/com/warehouse/warehouse/persistance/domain/ProductDao.kt
@@ -1,9 +1,10 @@
 package com.warehouse.warehouse.persistance.domain
 
+import java.math.BigInteger
 import java.time.LocalDateTime
 
 data class ProductDao(
-    val id: Long,
+    val id: BigInteger,
     val title: String?,
     val handle: String?,
     val productType: String?,
